Renaming namepsaces to be Addresses
