+++ AddressesDataPipeline.Tests/GlobalSuppressions.cs
@@ -5,4 +5,4 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-[assembly: SuppressMessage("Globalization", "CA1305:Specify IFormatProvider", Justification = "<Pending>", Scope = "member", Target = "~M:AddressDataPipeline.Tests.DatabaseTests.CountRows~System.Int64")]
+[assembly: SuppressMessage("Globalization", "CA1305:Specify IFormatProvider", Justification = "<Pending>", Scope = "member", Target = "~M:AddressesDataPipeline.Tests.DatabaseTests.CountRows~System.Int64")]
